refactor: modularize LineConfidenceChart by integrating hooks and components
- Replaced local state management and calculations with custom hooks for chart dimensions, scales, and brush functionality, enhancing code clarity and maintainability.
- Introduced ChartTooltip component for improved tooltip management, consolidating tooltip logic and enhancing user experience.
- Updated default props and removed unused code to streamline the component structure and improve readability.
