+++ src/app/LineConfidenceChart/components/ChartTooltip.tsx
@@ -1,0 +1,124 @@
+import React from 'react';
+import { AggregatedMeasurement, MeasurementItem } from '@upstream/upstream-api';
+import NumberFormatter from '../../common/NumberFormatter/NumberFormatter';
+
+// Types
+export interface TooltipData {
+  x: number;
+  y: number;
+  data: AggregatedMeasurement;
+}
+
+export interface PointTooltipData extends Partial<MeasurementItem> {
+  x: number;
+  y: number;
+}
+
+interface ChartTooltipProps {
+  tooltip: TooltipData | null;
+  tooltipPoint: PointTooltipData | null;
+  setTooltipAggregation: React.Dispatch<
+    React.SetStateAction<TooltipData | null>
+  >;
+  setTooltipPoint: React.Dispatch<
+    React.SetStateAction<PointTooltipData | null>
+  >;
+}
+
+const ChartTooltip: React.FC<ChartTooltipProps> = ({
+  tooltip,
+  tooltipPoint,
+  setTooltipAggregation,
+  setTooltipPoint,
+}) => {
+  return (
+    <>
+      {tooltip && (
+        <div
+          className="absolute bg-white border border-gray-300 rounded shadow-lg p-2 text-sm"
+          style={{
+            left: tooltip.x + 10,
+            top: tooltip.y - 10,
+            pointerEvents: 'auto',
+          }}
+        >
+          <div className="flex justify-between items-start mb-2">
+            <div className="font-medium">Aggregated Point Details</div>
+            <button
+              onClick={() => setTooltipAggregation(null)}
+              className="ml-2 text-gray-500 hover:text-gray-700"
+              aria-label="Close tooltip"
+            >
+              <svg
+                className="w-4 h-4"
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth={2}
+                  d="M6 18L18 6M6 6l12 12"
+                />
+              </svg>
+            </button>
+          </div>
+          <div>Time: {tooltip.data.measurementTime.toLocaleString()}</div>
+          <div>
+            Value: <NumberFormatter value={tooltip.data.value} />
+          </div>
+          <div>
+            Median: <NumberFormatter value={tooltip.data.medianValue} />
+          </div>
+          <div>
+            Confidence Interval: [
+            <NumberFormatter value={tooltip.data.parametricLowerBound} />,
+            <NumberFormatter value={tooltip.data.parametricUpperBound} /> ]
+          </div>
+          <div>Point Count: {tooltip.data.pointCount}</div>
+        </div>
+      )}
+
+      {tooltipPoint && (
+        <div
+          className="absolute bg-white border border-gray-300 rounded shadow-lg p-2 text-sm"
+          style={{
+            left: tooltipPoint.x + 10,
+            top: tooltipPoint.y - 10,
+            pointerEvents: 'auto',
+          }}
+        >
+          <div className="flex justify-between items-start mb-2">
+            <div className="font-medium">Individual Point Details</div>
+            <button
+              onClick={() => setTooltipPoint(null)}
+              className="ml-2 text-gray-500 hover:text-gray-700"
+              aria-label="Close tooltip"
+            >
+              <svg
+                className="w-4 h-4"
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth={2}
+                  d="M6 18L18 6M6 6l12 12"
+                />
+              </svg>
+            </button>
+          </div>
+          <div>Time: {tooltipPoint.collectiontime?.toLocaleString()}</div>
+          <div>
+            Value: <NumberFormatter value={tooltipPoint.value ?? 0} />
+          </div>
+        </div>
+      )}
+    </>
+  );
+};
+
+export default ChartTooltip;
